refactor: LottoResultStatus 패키지 변경
