+++ src/main/java/lotto/domain/Lotto.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
+import lotto.constant.LottoResultStatus;
 import lotto.exception.LottoExceptionMessage;
 
 public class Lotto {
